# Reduce allocations when solving path in `AStarGrid2D`