Improves path solving performance in `AStarGrid2D`.

| Test | Before | After |
|--|--|--|
| astar_grid_500_times_500_points | 5176ms | 2647ms |

Benchmark: [astar_grid_2d.gd.zip](https://github.com/user-attachments/files/17265994/astar_grid_2d.gd.zip)